@@ -2,45 +2,39 @@
   "name": "material-ui-forms-examples",
   "version": "0.1.0",
   "private": true,
+  "scripts": {
+    "dev": "next dev -p 3000",
+    "build": "next build",
+    "start": "next start"
+  },
   "dependencies": {
-    "@material-ui/core": "^4.11.3",
-    "@material-ui/icons": "^4.11.2",
-    "@material-ui/lab": "^4.0.0-alpha.57",
-    "@types/next": "^9.0.0",
-    "@types/react": "^17.0.1",
+    "@emotion/react": "^11.7.0",
+    "@emotion/styled": "^11.6.0",
+    "@fortawesome/fontawesome-svg-core": "^1.2.36",
+    "@fortawesome/free-brands-svg-icons": "^5.15.4",
+    "@fortawesome/free-solid-svg-icons": "^5.15.4",
+    "@fortawesome/react-fontawesome": "^0.1.16",
+    "@mui/icons-material": "^5.2.0",
+    "@mui/lab": "^5.0.0-alpha.57",
+    "@mui/material": "^5.2.1",
+    "@mui/styles": "^5.2.1",
     "axios": "^0.21.1",
-    "material-ui-forms": "file:..\\material-ui-forms-1.0.13.tgz",
-    "next": "^10.0.7",
-    "react": "^17.0.1",
-    "react-dom": "^17.0.1",
-    "react-to-print": "^2.12.3",
-    "typescript": "^4.2.2"
+    "clsx": "^1.1.1",
+    "date-fns": "^2.26.0",
+    "markdown-to-jsx": "^7.1.3",
+    "material-ui-forms": "file:..\\material-ui-forms-2.0.0.tgz",
+    "next": "^11.1.2",
+    "next-auth": "^4.0.0-beta.7",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-to-print": "^2.12.3"
   },
   "devDependencies": {
     "@types/axios": "^0.14.0",
-    "react-scripts": "^4.0.3"
-  },
-  "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "server": "NODE_ENV=production node server.js",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": "react-app"
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+    "@types/next": "^9.0.0",
+    "@types/next-auth": "^3.15.0",
+    "@types/react": "^17.0.3",
+    "cross-env": "^7.0.3",
+    "typescript": "^4.5.2"
   }
 }
