@@ -27,27 +27,34 @@
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
-  "version": "1.0.16",
+  "version": "2.0.0",
   "peerDependencies": {
-    "@material-ui/core": "4 - 5",
-    "@material-ui/icons": "4 - 5",
-    "@material-ui/lab": ">=4.0.0-alpha.50",
-    "next": "10 - 11",
-    "react": "17 - 18",
-    "react-dom": "17 - 18",
-    "react-to-print": "2 - 3"
+    "@emotion/react": ">=11.7.0",
+    "@emotion/styled": ">=11.6.0",
+    "@mui/icons-material": ">=5.2.0",
+    "@mui/lab": ">=5.0.0-alpha.57",
+    "@mui/material": ">=5.2.1",
+    "@mui/styles": ">=5.2.1",
+    "@types/crypto-js": ">=4.0.2",
+    "next": ">=11.1.2",
+    "react": ">=17.0.2",
+    "react-dom": ">=17.0.2",
+    "react-to-print": ">=2.12.3"
   },
   "devDependencies": {
-    "@material-ui/core": "4 - 5",
-    "@material-ui/icons": "4 - 5",
-    "@material-ui/lab": "4 - 5",
-    "@types/crypto-js": "4 - 5",
-    "next": "10 -11",
-    "react": "17 - 18",
-    "react-dom": "17 - 18",
-    "react-to-print": "2 - 3"
+    "@emotion/react": "^11.7.0",
+    "@emotion/styled": "^11.6.0",
+    "@mui/icons-material": "^5.2.0",
+    "@mui/lab": "^5.0.0-alpha.57",
+    "@mui/material": "^5.2.1",
+    "@mui/styles": "^5.2.1",
+    "@types/crypto-js": "^4.0.2",
+    "next": "^11.1.2",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
+    "react-to-print": "^2.12.3"
   },
   "dependencies": {
-    "crypto-js": "4 - 5"
+    "crypto-js": "^4.1.1"
   }
 }
