@@ -1,11 +1,11 @@
 import React from 'react';
-import Button from '@material-ui/core/Button';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import DialogTitle from '@material-ui/core/DialogTitle';
-import { PropTypes } from '@material-ui/core';
+import Button from '@mui/material/Button';
+import Dialog from '@mui/material/Dialog';
+import DialogActions from '@mui/material/DialogActions';
+import DialogContent from '@mui/material/DialogContent';
+import DialogContentText from '@mui/material/DialogContentText';
+import DialogTitle from '@mui/material/DialogTitle';
+import { PropTypes } from '@mui/material';
 
 //Todo Generate random UID if no id supplied
 
@@ -45,10 +45,10 @@ export function AlertDialog(props: IAlertDialogProps) {
         </DialogContentText>
       </DialogContent>
       <DialogActions>
-        <Button onClick={() => handleClose(false)} color={props.color ? props.color : "primary"}>
+        <Button onClick={() => handleClose(false)} sx={{ color: props.color ? props.color : "primary" }}>
           {props.backLabel || "Cancel"}
         </Button>
-        <Button onClick={() => handleClose(true)} color={props.color ? props.color : "primary"} autoFocus>
+        <Button onClick={() => handleClose(true)} sx={{ color: props.color ? props.color : "primary" }} autoFocus>
           {props?.forwardLabel || "Continue"}
         </Button>
       </DialogActions>
