@@ -1,5 +1,5 @@
 import React, { ChangeEvent, useState } from "react";
-import { Checkbox, FormControlLabel } from "@material-ui/core";
+import { Checkbox, FormControlLabel } from "@mui/material";
 
 export interface ICheckFieldProps {
     name: string;
