@@ -1,5 +1,5 @@
-import { SnackbarOrigin } from "@material-ui/core";
-import { AlertProps } from "@material-ui/lab/Alert";
+import { SnackbarOrigin } from "@mui/material";
+import { AlertProps } from '@mui/material/Alert';
 import { IRepoController } from "../interfaces";
 
 export type AlertState = {
