@@ -1,1 +1,1 @@
-self.__BUILD_MANIFEST = {__rewrites:[],"/":["static\u002Fchunks\u002Fpages\u002Findex.js"],"/_error":["static\u002Fchunks\u002Fpages\u002F_error.js"],"/views/form-view-auto":["static\u002Fchunks\u002Fpages\u002Fviews\u002Fform-view-auto.js"],sortedPages:["\u002F","\u002F_app","\u002F_error","\u002Fviews\u002Fform-view-auto"]};self.__BUILD_MANIFEST_CB && self.__BUILD_MANIFEST_CB()
+self.__BUILD_MANIFEST = {__rewrites:{beforeFiles:[],afterFiles:[],fallback:[]},"/":["static\u002Fchunks\u002Fpages\u002Findex.js"],"/_error":["static\u002Fchunks\u002Fpages\u002F_error.js"],"/next/dist/pages/_error":["static\u002Fchunks\u002Fpages\u002Fnext\u002Fdist\u002Fpages\u002F_error.js"],"/views/form-view":["static\u002Fchunks\u002Fpages\u002Fviews\u002Fform-view.js"],sortedPages:["\u002F","\u002F_app","\u002F_error","\u002Fnext\u002Fdist\u002Fpages\u002F_error","\u002Fviews\u002Fform-view"]};self.__BUILD_MANIFEST_CB && self.__BUILD_MANIFEST_CB()
